+++ src/db/index.ts
@@ -6,6 +6,7 @@ import City from './models/City'
 import ExplorerProfile from './models/ExplorerProfile'
 import Follow from './models/Follow'
 import PendingMedia from './models/PendingMedia'
+import PostDraft from './models/PostDraft'
 import Review from './models/Review'
 import Spot from './models/Spot'
 import SyncFailure from './models/SyncFailure'
@@ -15,6 +16,7 @@ export { default as City } from './models/City'
 export { default as ExplorerProfile } from './models/ExplorerProfile'
 export { default as Follow } from './models/Follow'
 export { default as PendingMedia } from './models/PendingMedia'
+export { default as PostDraft } from './models/PostDraft'
 export { default as Review } from './models/Review'
 export { default as Spot } from './models/Spot'
 export { default as SyncFailure } from './models/SyncFailure'
@@ -29,6 +31,7 @@ export const modelClasses = [
   City,
   SyncFailure,
   PendingMedia,
+  PostDraft,
 ]
 
 /**